Fixed errors on watch

+++ src/Game.h
@@ -6,8 +6,8 @@ typedef struct
 {
   int8_t gameInPlay;
   int game_time;
-  int score;
-  int highScore;
+  uint16_t score;
+  uint16_t highScore;
   int8_t speed;
   int8_t crash;
   int timeOfLastUpdate;
